+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Multiple direct message sending attempts
 * Press `Ctrl+Up` to insert last sent message into the text field
 * Add arrows to font
+* Fix screen not updated on new message
 
 # 2026.7.0